// src/common/interfaces/IColors.ts
export interface IColors {
    primary_50: string | null;
    primary_100: string | null;
    primary_200: string | null;
    primary_300: string | null;
    primary_400: string | null;
    primary_500: string | null;
    primary_600: string | null;
    primary_700: string | null;
    primary_800: string | null;
    primary_900: string | null;
}